oidc_login/requirements.txt
functions-framework==3.*
google-cloud-logging
google-cloud-bigtable
google-cloud-core
firebase-admin